Refactor and optimize code for the Cargo Target Configurations - Fixes #33 and #35. - Old APIs are removed and no longer supported
